src/backend/pyproject.toml
[project]
name = "gitulyse_api"
version = "0.0.1"
description = "A simple API for Gitulyse"
dependencies = [
    "flask",
    "pygithub",
    "pymongo",
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["gitulyse_api"]